cov-mockbuild: append :/ to $PATH used in chroot
... in order to be able to scan openjdk-1.8.0 with clang-analyzer.  Its
configure script (toolchain.m4 -> generated-configure.sh) tries to find
the compiler in the path in a very strange way because it concatenates
/usr/libexec/clang-analyzer/scan-build/ccc-analyzer with each single
item of $PATH and checks whether such a file exists.  Perhaps they do
not expect $CC to be an absolute path?  Anyway, appending :/ to $PATH
makes this bug disappear.