csmock: always clean mock chroot on exit
... unless --no-clean is given.

Previously csmock did not clean mock chroot if a non-recoverable failure
occurred before the chroot became unused.  It was useful for debugging
failures when running csmock interactively.

It is not so useful when running csmock by Covscan workers, which ran
out of disk space when more such failures occurred in multiple mock
chroots when analyzing packages like firefox or thunderbird.