csmock: apply filters of important reports more precisely
In order not to clobber scan properties, we used to use `csmock
--mode=grep` while feeding csdiff.  This caused spurious matches
of important reports in case csdiff parsed the plain-text output
differently (namely a wrong event was recognized as the key event
in some corner cases).  This commit makes csdiff operate solely
on the JSON format, which tags key events explicitly.