csmock: do not hard-wire ordering of plug-ins
Instead of assigning unique numbers to each plug-in, let them define
relative dependencies via the "pass_before" and "pass_after" attributes
of PluginProps.

The topological sorting is not yet fully implemented because we would
like to use graphlib.TopologicalSorter which is not available in
python-2, which we still care about.

Note that, when the order is otherwise unspecified, the plug-ins are
sorted alphabetically by their names.  This is an observable change
in behavior, compared to how it worked before.

Upstream PR: https://github.com/kdudka/csmock/pull/30