csmock: prevent `rpmbuild -bi` from failing on unpackaged files
This is a frequent cause of build/scan failures due to bugs in packaging
where redundant files are left in the buildroot when RPM packages are
built without tests or documentation.  New cases were also triggered by
the unicontrol plug-in which makes python3 unexpectedly available in el7
or el8 buildroots.  This fixes a scan failure for fetchmail-6.4.24-1.el8
with shellcheck and unicontrol plug-ins enabled:

error: Installed (but unpackaged) file(s) found:
   /usr/lib/python3.6/site-packages/__pycache__/fetchmailconf.cpython-36.opt-1.pyc
   /usr/lib/python3.6/site-packages/__pycache__/fetchmailconf.cpython-36.pyc
    Installed (but unpackaged) file(s) found:
   /usr/lib/python3.6/site-packages/__pycache__/fetchmailconf.cpython-36.opt-1.pyc
   /usr/lib/python3.6/site-packages/__pycache__/fetchmailconf.cpython-36.pyc