csmock: install optional pkgs one by one
If we run `mock --install csdiff compat-openssl11` and
`compat-openssl11` is not available, it results in a failed transaction
and none of them is installed.

If we run `mock --install` for each package separately, it will install
everything that is possible to install.  `mock --install --skip-broken`
would be more efficient but it is not available.