results: set the "imp" flag in the list of important findings
... rather than clearing it.  If scan results are processed later on,
it is useful to know the original state of the "imp" flag, regardless
of the context (whether we have a list of important findings only,
or an all-in-one list of findings).  A side effect will be that a red
`[important]` tag will appear in the HTML output next to each finding,
which is probably harmless.

Related: https://issues.redhat.com/browse/OSH-343
Related: https://issues.redhat.com/browse/OSH-565
Closes: https://github.com/csutils/csmock/pull/159