clippy: do not fail the scan if clippy fails to install
If the `clippy` package is not available in the build repos, a warning
is emitted and `clippy` is not recorded as an enabled tool in the scan
properties.  But the scan continues without `clippy` in this case.

The approach is similar to what `--gcc-analyze` does.  This is needed
because `clippy` is not available for each mock config and we do not
want to configure the set of enabled tools separately for each mock
config.

Related: https://issues.redhat.com/browse/OSH-30
Closes: https://github.com/csutils/csmock/pull/174