csmock --kfp-git-url: direct access to a KFP repo
If this option is used, one does not need to build and install the
`known-false-positives` RPM package.  The KFP data is read directly
from the specified git repository.

This feature internally uses the `csfilter-kfp` utility from csdiff
and it works independently of the `--known-false-positives` option,
which is kept for backward compatibility.

Resolves: https://github.com/csutils/csmock/issues/201
PR: https://github.com/csutils/csmock/pull/203